Chunked writing to hdf
Changed the HDF5 writing to be performed in chunks, and to display a progress bar.
Helpful for larger line lists, such as SiH4, to seee the progress of the HDF5 conversion.
Added one helper method for this. Also currently use the lzf compression scheme
when writing to hdf due to its speed.